+++ src/main/java/com/revature/models/Product.java
@@ -21,7 +21,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer productId;
+    private long productId;
     private String name;
     private String description;
     private Integer quantity;
